+++ internal/repository/galaxy.go
@@ -7,27 +7,24 @@ import (
 	"cosmos/internal/models"
 )
 
-// GalaxyRepository - интерфейс для работы с галактиками
 type GalaxyRepository interface {
 	GetAll() ([]models.Galaxy, error)
-	GetByID(id int) (*models.Galaxy, error)
+	GetByID(id int64) (*models.Galaxy, error) // int → int64
 	Create(galaxy *models.Galaxy) error
 	Update(galaxy *models.Galaxy) error
-	Delete(id int) error
+	Delete(id int64) error // int → int64
 	Count() (int, error)
-	GetPlanetCount(id int) (int, error) // проверка есть ли планеты
+	GetPlanetCount(id int64) (int, error) // int → int64
 }
 
 type galaxyRepository struct {
 	db *sql.DB
 }
 
-// NewGalaxyRepository - конструктор
 func NewGalaxyRepository(db *sql.DB) GalaxyRepository {
 	return &galaxyRepository{db: db}
 }
 
-// GetAll - получить все галактики
 func (r *galaxyRepository) GetAll() ([]models.Galaxy, error) {
 	query := `
 		SELECT id, name, type, diameter_ly, mass_suns,
@@ -79,8 +76,7 @@ func (r *galaxyRepository) GetAll() ([]models.Galaxy, error) {
 	return galaxies, nil
 }
 
-// GetByID - получить галактику по ID
-func (r *galaxyRepository) GetByID(id int) (*models.Galaxy, error) {
+func (r *galaxyRepository) GetByID(id int64) (*models.Galaxy, error) { // int64
 	var galaxy models.Galaxy
 	var diameterLy, massSuns, distanceFromEarthLy sql.NullFloat64
 	var discoveredYear sql.NullInt64
@@ -124,7 +120,6 @@ func (r *galaxyRepository) GetByID(id int) (*models.Galaxy, error) {
 	return &galaxy, nil
 }
 
-// Create - создание галактики
 func (r *galaxyRepository) Create(galaxy *models.Galaxy) error {
 	query := `
 		INSERT INTO galaxies (name, type, description, diameter_ly, mass_suns,
@@ -165,7 +160,6 @@ func (r *galaxyRepository) Create(galaxy *models.Galaxy) error {
 	).Scan(&galaxy.ID, &galaxy.CreatedAt)
 }
 
-// Update - обновление галактики
 func (r *galaxyRepository) Update(galaxy *models.Galaxy) error {
 	query := `
 		UPDATE galaxies
@@ -219,8 +213,7 @@ func (r *galaxyRepository) Update(galaxy *models.Galaxy) error {
 	return nil
 }
 
-// Delete - удаление галактики
-func (r *galaxyRepository) Delete(id int) error {
+func (r *galaxyRepository) Delete(id int64) error { // int64
 	result, err := r.db.Exec("DELETE FROM galaxies WHERE id = $1", id)
 	if err != nil {
 		return err
@@ -234,15 +227,13 @@ func (r *galaxyRepository) Delete(id int) error {
 	return nil
 }
 
-// Count - количество галактик
 func (r *galaxyRepository) Count() (int, error) {
 	var count int
 	err := r.db.QueryRow("SELECT COUNT(*) FROM galaxies").Scan(&count)
 	return count, err
 }
 
-// GetPlanetCount - количество планет в галактике
-func (r *galaxyRepository) GetPlanetCount(id int) (int, error) {
+func (r *galaxyRepository) GetPlanetCount(id int64) (int, error) { // int64
 	var count int
 	err := r.db.QueryRow("SELECT COUNT(*) FROM planets WHERE galaxy_id = $1", id).Scan(&count)
 	return count, err
