+++ internal/service/planet_service.go
@@ -19,22 +19,18 @@ func NewPlanetService(planetRepo repository.PlanetRepository, galaxyRepo reposit
 	}
 }
 
-// GetAllPlanets - получить все планеты
 func (s *PlanetService) GetAllPlanets() ([]models.Planet, error) {
 	return s.planetRepo.GetAll()
 }
 
-// GetPlanetByID - получить планету по ID
-func (s *PlanetService) GetPlanetByID(id int) (*models.Planet, error) {
+func (s *PlanetService) GetPlanetByID(id int64) (*models.Planet, error) { // int64
 	if id <= 0 {
 		return nil, errors.New("неверный ID планеты")
 	}
 	return s.planetRepo.GetByID(id)
 }
 
-// CreatePlanet - создание планеты с валидацией
 func (s *PlanetService) CreatePlanet(planet *models.Planet) error {
-	// Валидация
 	if planet.Name == "" {
 		return errors.New("название планеты обязательно")
 	}
@@ -51,7 +47,6 @@ func (s *PlanetService) CreatePlanet(planet *models.Planet) error {
 	return s.planetRepo.Create(planet)
 }
 
-// UpdatePlanet - обновление планеты
 func (s *PlanetService) UpdatePlanet(planet *models.Planet) error {
 	if planet.ID <= 0 {
 		return errors.New("неверный ID планеты")
@@ -69,20 +64,17 @@ func (s *PlanetService) UpdatePlanet(planet *models.Planet) error {
 	return s.planetRepo.Update(planet)
 }
 
-// DeletePlanet - удаление планеты
-func (s *PlanetService) DeletePlanet(id int) error {
+func (s *PlanetService) DeletePlanet(id int64) error { // int64
 	if id <= 0 {
 		return errors.New("неверный ID планеты")
 	}
 	return s.planetRepo.Delete(id)
 }
 
-// GetPlanetCount - количество планет
 func (s *PlanetService) GetPlanetCount() (int, error) {
 	return s.planetRepo.Count()
 }
 
-// GetGalaxiesForSelect - получить галактики для выпадающего списка
 func (s *PlanetService) GetGalaxiesForSelect() ([]models.Galaxy, error) {
 	return s.planetRepo.GetGalaxies()
 }
